requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
sklearn
pyyaml
joblib